finished word guessing logic [meesterproef]
The logic where a user guesses a word is finished!

I ran into some issues when i was comparing the user guessed word with the random word. Like for example in the word woo, both o's being yellow marked, while the random word, wow, only has 1 o.

This would be really confusing for the user, so i made sure that the letters in the user guessed word, were only yellow market if they didn't exceed the number of occurences in the random word.

I did this with the help of 2 dictionaries, one for the amount of reacurring letters in the random word (with the help of the counter attribute, and dict, to make sure the layout of the word was still correct)

And a dict for the user guessed word letters, each letter has a 0 assigned too it, and every time the letter loop is initialized, the selected letter gets added by 1

Then i added an if statement with a try catch too see if the letter count in the guessed word dict exceeded the letter count in the random word dict, if so, a continue is executed.

The reason this if is in a try catch is because it is not always certain set letter is contained in the random word dict, hense why if a keyError accurs, a continue attribute is executed which restarts the loop.