new mastermind class [meesterproef]
decided too put the word guessing logic in a seperate class, this helps me keep things organized and easier to maintain as the project grows. And in the mastermind class, i also splitted up different logics like selecting a random word into methods.

this helps the class become easier to read and maintain :)

with the mastermind_game method in the mastermind class, a random word is selected, the duplicate letter list is made to keep track of duplicate letters. The visual word guesser is set up with the new prepare_word() method, and the word_guessing_logic() is executed