small feature: Team loses after not guessing a word 3 times in a row
I forgot too apply an additional rule too the game. If a team fails to guess a word 3 times in a row (for every new word, not try) they instantly lose and the other team wins.

Team.py:

added a new field to keep track of the falsly guessed words

mastermind.py:

Made sure the team field for the amount of falsly guessed words gets added up by 1 if a team fails to guess a word within 5 tries

lingo_game.py:

Added 2 new if's to keep track of the amountOfFailedGuessedWords field of both team 1 and team 2, if one of them exceeds 3, they lose and the other team wins.