New class: Basket [meesterproef]
This new class will provide the game with a new feature, the basket. If a team guesses a word correctly, they are allowed too grab a ball from their basket! (each time has their own basket). The actual functionality and what the balls are for will be developed and thought out later on