applying the new basket feature too the lingo_game class [meesterproef]
Too ensure a team can select a random ball from their basket after they have correctly guessed a word, i first needed to create a basket instance (line 16). Here i take the 2 teams and create their baskets. Team 1 has 3 green balls, 3 red balls and 17 odd numbers, Team 2 has the same, with the only difference being that they have even numbers instead of odd.

after that, i make sure that the baskets are created when the start_game() method is executed. Then i have 2 variables in the main loop that keep track of the current number of guessed words, this is needed for the following 2 new if's

these new if's (line 59 and 49) will keep track of any changes too the amountOfGuessedWords() field, if the if notices that the amountOfGuessedWords has increased, the randomBallfromBasket() method is executed