new fields and methods Team class [meesterproef]
I've created 2 new fields and 1 new method for the team class.

New fields:

- basket:
This will save the basket of the team

- ballsInPossession:
This fields keeps track of all the balls a team has in possession

New methods:

- RandomBallFromBasket():
this new method selects a random ball from the basket field, prints what ball the team has gotten, and appends it too the ballsInPossession field.