finished the basket class [meesterproef]
The basket class has been finished! Now when a team has 3 green balls, they instantly win. But if they have 3 red balls, the other team wins.

I might use an enum list later on since i do use a couple of magic strings

new methods:

- CheckPossibleWinnerBalls():

this method is used to determine the winner of the game, based on the amount of green or red balls a team has.

I take the duplicates of both the team1 and team2 basket. And then check the amount of red and green balls, to see if a winner can be determined, if not, the function will return false.

I also have these if statements in a try, except since it isnt always certain a team may have a green or red ball

- BasketRound():

this function is used for when a team can take a ball from their basket. they get a total of 2 turns, if they get a red ball on the first pull, they need to skip their second pull.

- TakeBallFromBasket():

this method simply takes a ball from the basket, puts it in the teams ballsInPossession list, and removes it from the basket (too prevent pulling duplicate items)