writting demo for new bingo class [meesterproef]
Before applying the new bingo feature directly too the lingo game, i decided to first make a sort of test demo, in order to see how i'll make the bingo class

I decided too go for an approach where both teams get their own bingo card at the beginning of the game, there is a bingo when there are 4 x's next to each other horizontally or vertically. (the numbers on the bingo card turn into an x if one of the numbered balls in the teams possession matches one of the numbers balls on the bingo card)

only thing left to make is the possible bingos a team can have. After that's sorted, i'll begin working on the bingo class.