finished bingo logic [meesterproef]
I finished the bingo logic for when a team has 4 x's in a row vertically and diagonally.

Vertical bingo:

I used two loops for this one:

the first loop is for looping through all the vertical columns, while the second loop is used to move from i row too the other.

so i use the for loop through the first vertically column, and that the second loop within the first loop, is used to move on too the next vertical column. (this took me a while to program since it is kind of like reverse logic)

Diagonal bingo:

for the diagonal bingo, i simply use the i of the for loop for both looping through the list AND The values, this means i get index values like (1, 1), (2, 2,), (3, 3). Which makes a diagonal line through the bingo card.

I also had to make an additional loop for the diagonal bingo logic, since the first one is only for diagonal bingos from left to right of the card, also needed logic for if a team had a diagonal bingo from right to left.

I did this by dividing i from 3 (highest index of the list) to sort of countdown through the list. So instead of (1, 1) (2, 2) (3, 3) (4, 4), you get (1, 4) (2, 3) (3, 2) etc.