new bingo class [meesterproef]
created the new bingo class for the new bingo feature of the game

[methods]:

- CreateBingoCard():
this method creates, like the name suggests, a bingo card, a team paramater is given in order to create the bingo card, and immidiately append it too the bingo_card field.

- CheckForMatchingNums():
this method is used to check out numbers that match the numbered balls in the teams possession.

- ShowBingoCard():
Like the name already suggests, this method is used to display the nicely printed bingo card of a team

- CheckForBingo() [not finished!]:

this method is used to determine wether or not a team has a bingo, like in the test file, it checks for 4 x's horizontally, diagonally and vertically. However, it seems like this method is not functioning correctly yet.

I'll fix the issue on a later data.