applying new bingo feature too lingo_game [meesterproef]
[instances]:

- CreateBingoCards():
for creating the bingo cards for both teams

- ShowBingoCards():
for printing out the bingo card of a given team

- CheckForBingo():
Too check whether or not a team got bingo.

[how it works]

when the game starts, both teams get their assigned bingo card, whenever they get 2 balls (or 1) and it matches one of the numbers on the bingo card, set number on the bingo card will be crossed out with an x. Whoever has 4 x's next to each other horizontally, diagonally or vertically, wins.

This feature is sadly not fully functional yet, most of the bingo logic is for debugging purposes.

Thankfully only the part for checking bingos is not functional, the rest is completely functional!