BUG fix [meesterproef]
the 3 incorrectly guessed words in a row loss didn't work properly. For some reason it only counted if a team had 3 incorrectly guessed words, not 3 incorrectly guessed words in a row.

which is why i just reset the amountOfFailedGuessedWords when the team guesses a word, this way, they will only lose if they incorrectly guess 3 words in a row.